requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.4
django-allauth==0.34.0
django-jwt-auth==0.0.2
django-node==4.0.0
django-rest-auth==0.9.2
django-rest-framework==0.1.0
django-templated-mail==1.0.0
django-webpack-loader==0.5.0
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2
djoser==1.1.5
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.0.6
omdb==0.8.1
optional-django==0.1.0
psycopg2==2.7.3.2
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
uritemplate==3.0.0
urllib3==1.22
